Re-implement VM list and VM information fetching
When fetching VM lists, all VMs and their corresponding information
all the VMs and complete information about each one of them was
fetched. This involved multiple API calls and loops to be executed
before the response can be returned to user.
This commit changes the code so as to fetch only the currently
required information and deffer other information for later.